3.3.0: vendor core v3.3.0 and widen channel-table bounds arithmetic
Vendors SuperFAISS core at commit e7323f4 (tag v3.3.0), which adds
PeekScratchArchive, validates a loaded scratch bank's retention region,
bounds Create/Grow geometry before the arena size is computed, and resolves
one float32 kernel-selection rule so the segmented and whole-row scans agree.
ComputePrincipalComponents' scratch parameter became double*, so this plugin's
one call site in SSuperFAISSBankInspector::ComputeProjection changes its
buffer from TArray<float> to TArray<double>.

Widens the caller-supplied channel Offset + Length sum to int64 before every
bounds comparison, in InitFromSource, RebuildChannelTable, InitWithChannels
and Relabel. Two large positive int32 values could overflow before comparison,
defeating the channel's own bounds and ascending-order checks. A malformed
table was never accepted -- the core's ValidateBank rejects it downstream --
but the plugin's own guard was inert for that input, so the caller received a
generic downstream failure instead of the specific diagnosis, and the signed
overflow was undefined behavior regardless.

Also stops tracking Python bytecode produced by the ci/ checks.

+++ SuperFAISSUnreal/Source/SuperFAISSUnreal/Private/SuperFAISSScratchBank.cpp
@@ -168,7 +168,10 @@ bool USuperFAISSScratchBank::InitWithChannels(int32 Capacity, int32 Dims,
 	{
 		superfaiss::ChannelInfo Info;
 		Info.offset = InChannelOffsets[C];
-		Info.length = (InChannelOffsets[C] + InChannelLengths[C] == Dims)
+		// Widen to int64 before comparing (F4, D-V32-89): Offset and Length are
+		// caller-supplied int32s, and their sum can overflow before it is compared
+		// against Dims.
+		Info.length = (static_cast<int64>(InChannelOffsets[C]) + InChannelLengths[C] == Dims)
 			? PaddedDims - InChannelOffsets[C]
 			: InChannelLengths[C];
 		Table.Add(Info);
@@ -230,7 +233,10 @@ bool USuperFAISSScratchBank::Relabel(const TArray<FName>& InChannelNames,
 	{
 		superfaiss::ChannelInfo Info;
 		Info.offset = InChannelOffsets[C];
-		Info.length = (InChannelOffsets[C] + InChannelLengths[C] == Dims)
+		// Widen to int64 before comparing (F4, D-V32-89): Offset and Length are
+		// caller-supplied int32s, and their sum can overflow before it is compared
+		// against Dims.
+		Info.length = (static_cast<int64>(InChannelOffsets[C]) + InChannelLengths[C] == Dims)
 			? PaddedDims - InChannelOffsets[C]
 			: InChannelLengths[C];
 		Table.Add(Info);
